間違った宣言を修正

+++ q10251743340/6-4.h
@@ -1,6 +1,6 @@
 extern int MIN3;
 extern int MAX3;
-extern int AVE_MIN_MAX;
+extern double AVE_MIN_MAX;
 
 void min3(int a, int b, int c);
 void max3(int a, int b, int c);
